requirements.txt
streamlit==1.28.0
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.24.3
pandas==1.5.3
sentence-transformers==2.2.2
transformers==4.33.0
torch==2.1.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
streamlit-chat==0.1.1